requirements.txt
pika
pyncclient
limesurveyrc2api
python-dotenv
fpdf2